run-with-google-adk/agents/google_mcp_security_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.97.0
google-adk[eval]>=1.5.0,<1.6.0
google-genai>=1.24.0
markdown
pandas
requests
python-dotenv
pydantic==2.11.7
cloudpickle==3.1.1